refactor: inline logic in PageLayoutRenderer
