+++ packages/twenty-front/src/modules/page-layout/components/PageLayoutRenderer.tsx
@@ -1,49 +1,50 @@
 import { PageLayoutInitializationQueryEffect } from '@/page-layout/components/PageLayoutInitializationQueryEffect';
 
 import { PageLayoutRendererContent } from '@/page-layout/components/PageLayoutRendererContent';
+import { useSetIsPageLayoutInEditMode } from '@/page-layout/hooks/useSetIsPageLayoutInEditMode';
 import { PageLayoutComponentInstanceContext } from '@/page-layout/states/contexts/PageLayoutComponentInstanceContext';
 import { type PageLayout } from '@/page-layout/types/PageLayout';
 import { getTabListInstanceIdFromPageLayoutId } from '@/page-layout/utils/getTabListInstanceIdFromPageLayoutId';
-import { LayoutRenderingProvider } from '@/ui/layout/contexts/LayoutRenderingContext';
+import { isPageLayoutEmpty } from '@/page-layout/utils/isPageLayoutEmpty';
 import { TabListComponentInstanceContext } from '@/ui/layout/tab-list/states/contexts/TabListComponentInstanceContext';
 import 'react-grid-layout/css/styles.css';
 import 'react-resizable/css/styles.css';
-import { PageLayoutType } from '~/generated/graphql';
 
 type PageLayoutRendererProps = {
   pageLayoutId: string;
-  onInitialized?: (pageLayout: PageLayout) => void;
 };
 
 export const PageLayoutRenderer = ({
   pageLayoutId,
-  onInitialized,
 }: PageLayoutRendererProps) => {
+  const { setIsPageLayoutInEditMode } =
+    useSetIsPageLayoutInEditMode(pageLayoutId);
+
+  const onInitialized = (pageLayout: PageLayout) => {
+    if (isPageLayoutEmpty(pageLayout)) {
+      setIsPageLayoutInEditMode(true);
+    } else {
+      setIsPageLayoutInEditMode(false);
+    }
+  };
+
   return (
-    <LayoutRenderingProvider
+    <PageLayoutComponentInstanceContext.Provider
       value={{
-        targetRecordIdentifier: undefined,
-        layoutType: PageLayoutType.DASHBOARD,
-        isInRightDrawer: false,
+        instanceId: pageLayoutId,
       }}
     >
-      <PageLayoutComponentInstanceContext.Provider
+      <TabListComponentInstanceContext.Provider
         value={{
-          instanceId: pageLayoutId,
+          instanceId: getTabListInstanceIdFromPageLayoutId(pageLayoutId),
         }}
       >
-        <TabListComponentInstanceContext.Provider
-          value={{
-            instanceId: getTabListInstanceIdFromPageLayoutId(pageLayoutId),
-          }}
-        >
-          <PageLayoutInitializationQueryEffect
-            pageLayoutId={pageLayoutId}
-            onInitialized={onInitialized}
-          />
-          <PageLayoutRendererContent />
-        </TabListComponentInstanceContext.Provider>
-      </PageLayoutComponentInstanceContext.Provider>
-    </LayoutRenderingProvider>
+        <PageLayoutInitializationQueryEffect
+          pageLayoutId={pageLayoutId}
+          onInitialized={onInitialized}
+        />
+        <PageLayoutRendererContent />
+      </TabListComponentInstanceContext.Provider>
+    </PageLayoutComponentInstanceContext.Provider>
   );
 };
